requirements.txt
feedparser==6.0.8
requests==2.26.0
PyYAML==5.4.1
python-dateutil==2.8.2 